requirements.txt
absl-py==2.3.1
astunparse==1.6.3
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
easyocr==1.7.2
filelock==3.13.1
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.59.1
fsspec==2024.6.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.74.0
h5py==3.14.0
idna==3.7
imageio==2.37.0
Jinja2==3.1.4
keras==3.11.2
kiwisolver==1.4.9
lazy_loader==0.4
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mdurl==0.1.2
ml_dtypes==0.5.3
MouseInfo==0.1.3
mpmath==1.3.0
mss==10.0.0
namex==0.1.0
networkx==3.3
ninja==1.13.0
numpy==2.2.6
opencv-contrib-python==4.12.0.88
opencv-python-headless==4.12.0.88
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.1
pi_heif==1.1.0
pillow==11.3.0
pillow-avif-plugin==1.5.2
protobuf==6.32.0
psutil==7.0.0
py-cpuinfo==9.0.0
PyAutoGUI==0.9.53
pyclipper==1.3.0.post6
PyGetWindow==0.0.9
Pygments==2.19.2
PyMsgBox==1.0.9
pynput==1.8.1
pyparsing==3.2.3
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==0.1.28
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytweening==1.2.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.1.0
roboflow==1.2.6
scikit-image==0.25.2
scipy==1.16.1
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
sympy==1.13.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow-intel==2.16.1
termcolor==3.1.0
tifffile==2025.6.11
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.179
ultralytics-thop==2.0.15
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
